Version 1.2.2.

+++ tools/ocr.py
@@ -1,7 +1,7 @@
 from PIL import Image
 from config import settings
 from colorama import Fore, Style
-import pytesseract, os, logging
+import pytesseract, os, logging, uuid
 
 logging.basicConfig(
     filename=settings.LOG_FILE,
@@ -28,9 +28,7 @@ class OCRProcessor:
     async def extract_text(self, image_path):
         """
         Extrait le texte d'une image donnée.
-        
         :param image_path: Chemin de l'image à analyser.
-        :return: Texte extrait de l'image.
         """
         try:
             # Ouvrir l'image
@@ -47,28 +45,34 @@ class OCRProcessor:
 
     async def process_attachment(self, attachment):
         """
-        Télécharge et analyse une pièce jointe Discord.
-        
+        Télécharge et analyse une pièce jointe Discord.     
         :param attachment: Pièce jointe Discord.
-        :return: Texte extrait de l'image ou message d'erreur.
         """
         try:
             # Télécharger l'image localement
-            local_filename = os.path.join(settings.UPLOAD_IMAGES, attachment.filename)
+            local_filename = os.path.join(settings.UPLOAD_IMAGES, f"{uuid.uuid4()}_{attachment.filename}")
             await attachment.save(local_filename)
             print(Fore.CYAN + f"[INFO] Image téléchargée : {local_filename}" + Style.RESET_ALL)
             self.logger.info(f"Image téléchargée : {local_filename}")
 
             # Extraire le texte de l'image
             extracted_text = await self.extract_text(local_filename)
-    
-            # Supprimer le fichier local après traitement
-            os.remove(local_filename)
-            print(Fore.YELLOW + f"[INFO] Fichier supprimé : {local_filename}" + Style.RESET_ALL)
-            self.logger.info(f"Fichier supprimé : {local_filename}")
-            
-            return extracted_text if extracted_text else "❌ Aucun texte détecté dans l'image."
+            return "Description de l'image:" + extracted_text if extracted_text else "❌ Aucun texte détecté dans l'image."
+        
         except Exception as e:
             print(Fore.RED + f"[ERROR] Erreur lors du traitement de la pièce jointe : {e}" + Style.RESET_ALL)
             self.logger.error(f"Erreur lors du traitement de la pièce jointe : {e}")
             return f"❌ Une erreur s'est produite lors de l'analyse de l'image : {e}"
+        finally:
+            if os.path.exists(local_filename):
+                try:
+                    # Supprimer le fichier local après traitement
+                    os.remove(local_filename)
+                    print(Fore.YELLOW + f"[INFO] Fichier supprimé : {local_filename}" + Style.RESET_ALL)
+                    self.logger.info(f"Fichier supprimé : {local_filename}")
+                except Exception as e:
+                    print(Fore.RED + f"[ERROR] Erreur lors de la suppression du fichier {local_filename} : {e}" + Style.RESET_ALL)
+                    self.logger.error(f"Erreur lors de la suppression du fichier {local_filename} : {e}")
+            
+            
+
